Implement Review comments
