+++ docs/contributor/04-local-test-setup.md
@@ -34,7 +34,7 @@ This setup is deployed with the following security features enabled:
     ```shell
     k3d cluster create kcp-local --port 9443:443@loadbalancer \
     --registry-create k3d-registry.localhost:0.0.0.0:5111 \
-    --k3s-arg '--disable=traefik@server:0'
+    --k3s-arg '--disable=traefik@server:0' \
     --k3s-arg --tls-san=host.k3d.internal@server:* 
     ```
