+++ internal/controller/kyma/setup.go
@@ -22,8 +22,8 @@ import (
 
 	"github.com/kyma-project/lifecycle-manager/api/shared"
 	"github.com/kyma-project/lifecycle-manager/api/v1beta2"
+	"github.com/kyma-project/lifecycle-manager/internal/watch"
 	"github.com/kyma-project/lifecycle-manager/pkg/security"
-	"github.com/kyma-project/lifecycle-manager/pkg/watch"
 )
 
 type SetupOptions struct {
