+++ internal/watch/mandatory_template_change.go
